Use budget period rather than separate endpoint
