+++ BudgetSquirrel.Api/ClientApp/src/app/budgets/services/budget-api.service.ts
@@ -27,7 +27,10 @@ export class BudgetApi {
 
   public getRootBudget(): Observable<Budget> {
     return this.http.get(`${this.baseUrl}${BUDGETS_API}/root-budget`).pipe(
-      map((budget: Budget) => budget)
+      map((budget: Budget) => {
+        budget.dateFinalized = new Date(budget.dateFinalized);
+        return budget;
+      })
     );
   }
