+++ BudgetSquirrel.Api/ClientApp/src/app/shared/models/core.ts
