--- BudgetSquirrel.Api/ClientApp/src/app/shared/services/tracking.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { TrackingService } from './tracking.service';
-
-describe('TrackingServiceService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: TrackingService = TestBed.get(TrackingService);
-    expect(service).toBeTruthy();
-  });
-});
